refactor(app): standardize benchmark reporting
